requirements.txt
django==4.2.18
python-telegram-bot==21.10